fixed bug on hhn and hn

+++ libft/ft_iterative_power_longlong.c
@@ -6,14 +6,14 @@
 /*   By: rmartins <rmartins@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/12/03 10:05:37 by rmartins          #+#    #+#             */
-/*   Updated: 2021/03/03 16:39:51 by rmartins         ###   ########.fr       */
+/*   Updated: 2021/03/06 22:29:07 by rmartins         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 unsigned long long	ft_iterative_power_ll(unsigned long long nb, int power)
 {
-	int	i;
-	int number;
+	int					i;
+	unsigned long long	number;
 
 	number = nb;
 	if (power < 0)
